+++ src/native_class.rs
@@ -37,7 +37,7 @@ pub fn get_primitive_class(frame: &mut StackFrame) {
                                     _ => panic!(),
                                 }
                             }
-                            let class_object: u32;
+                            let class_object: u64;
                             if v == vec!['f', 'l', 'o', 'a', 't'] {
                                 class_object =
                                     create_class_object(&String::from("java/lang/Float"));
